Move reconcileVizDataset to a common utils directory
